pyproject.toml
[tool.poetry]
name = "duty-bot"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "python_duty_bot"}]

[tool.poetry.dependencies]
python = "~3.9"
python-telegram-bot = {extras = ["job-queue"], version = "^20.2"}
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
